BallByBall get api updated with method asperrequirement

+++ src/main/java/sportify/backend/api/util/JavaApiClass/iplBallByBall/Score.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 public class Score {
     private int r;
     private int w;
-    private int o;
+    private float o;
     private String inning;
 }
